+++ src/main/java/betterquesting/client/gui2/editors/GuiTextEditor.java
@@ -19,7 +19,6 @@ import org.lwjgl.input.Keyboard;
 
 import betterquesting.api.client.gui.misc.IVolatileScreen;
 import betterquesting.api.misc.ICallback;
-import betterquesting.api.utils.RenderUtils;
 import betterquesting.api2.client.gui.GuiScreenCanvas;
 import betterquesting.api2.client.gui.controls.IPanelButton;
 import betterquesting.api2.client.gui.controls.PanelButton;
@@ -46,6 +45,7 @@ import betterquesting.api2.client.gui.popups.PopWaitExternalEvent;
 import betterquesting.api2.client.gui.themes.presets.PresetColor;
 import betterquesting.api2.client.gui.themes.presets.PresetTexture;
 import betterquesting.api2.utils.QuestTranslation;
+import betterquesting.api2.utils.TextFormattingUtils;
 import betterquesting.core.BetterQuesting;
 import betterquesting.misc.QuestResourcesFile;
 import betterquesting.misc.QuestResourcesFolder;
@@ -224,87 +224,8 @@ public class GuiTextEditor extends GuiScreenCanvas implements IPEventListener, I
         flText.writeText(text);
     }
 
-    private static boolean isHex6(String str, int start) {
-        if (start + 6 > str.length()) return false;
-        for (int k = 0; k < 6; k++) {
-            char c = str.charAt(start + k);
-            if (!((c >= '0' && c <= '9') || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F'))) return false;
-        }
-        return true;
-    }
-
     public static String stripFormatting(String text) {
-        StringBuilder sb = new StringBuilder();
-        int len = text.length();
-        for (int i = 0; i < len; i++) {
-            char ch = text.charAt(i);
-            if (ch == '&' && i + 1 < len) {
-                char next = text.charAt(i + 1);
-                char nextL = Character.toLowerCase(next);
-                // &g&#RRGGBB&#RRGGBB (18 chars)
-                if (nextL == 'g' && i + 18 <= len
-                    && text.charAt(i + 2) == '&'
-                    && text.charAt(i + 3) == '#'
-                    && isHex6(text, i + 4)
-                    && text.charAt(i + 10) == '&'
-                    && text.charAt(i + 11) == '#'
-                    && isHex6(text, i + 12)) {
-                    i += 17;
-                    continue;
-                }
-                // &#RRGGBB (8 chars)
-                if (next == '#' && isHex6(text, i + 2)) {
-                    i += 7;
-                    continue;
-                }
-                // &X single codes
-                if ((nextL >= '0' && nextL <= '9') || (nextL >= 'a' && nextL <= 'f')
-                    || (nextL >= 'k' && nextL <= 'o')
-                    || nextL == 'r'
-                    || nextL == 'x'
-                    || nextL == 'q'
-                    || nextL == 'z'
-                    || nextL == 'v'
-                    || nextL == 'g') {
-                    i += 1;
-                    continue;
-                }
-                // Literal &
-                sb.append(ch);
-            } else if (ch == '\u00a7' && i + 1 < len) {
-                char next = text.charAt(i + 1);
-                char nextL = Character.toLowerCase(next);
-                // §g + two §x sequences (30 chars)
-                if (nextL == 'g' && i + 30 <= len
-                    && RenderUtils.isValidSectionX(text, i + 2)
-                    && RenderUtils.isValidSectionX(text, i + 16)) {
-                    i += 29;
-                    continue;
-                }
-                // §x§R§R§G§G§B§B (14 chars)
-                if (nextL == 'x' && RenderUtils.isValidSectionX(text, i)) {
-                    i += 13;
-                    continue;
-                }
-                // §X (2 chars) — any known format code
-                if ((nextL >= '0' && nextL <= '9') || (nextL >= 'a' && nextL <= 'f')
-                    || (nextL >= 'k' && nextL <= 'o')
-                    || nextL == 'r'
-                    || nextL == 'x'
-                    || nextL == 'q'
-                    || nextL == 'z'
-                    || nextL == 'v'
-                    || nextL == 'g') {
-                    i += 1;
-                    continue;
-                }
-                // Unknown §, pass through
-                sb.append(ch);
-            } else {
-                sb.append(ch);
-            }
-        }
-        return sb.toString();
+        return TextFormattingUtils.stripFormatting(text);
     }
 
     private void writeImageTag(String resourceLoc) {
